requirements.txt
awscli==1.27.53
awscrt==0.15.3
boto3==1.26.52
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0